requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
idna==2.10
logzero==1.6.3
pydantic==1.7.3
pymongo==3.11.2
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
starlette==0.13.6
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
